grantedai/gcloud/requirements.txt
openai
bs4
requests
tiktoken
google-cloud-secret-manager==2.10.0
PyYAML==6.0